Fixed compiler errors

+++ Firmware/FFBoard/Src/CmdParser.cpp
@@ -156,7 +156,7 @@ bool CmdParser::parse(std::vector<ParsedCommand>& commands){
 						bool validPqm = (pqm != std::string::npos && (std::isdigit(word[pqm+1]) || (std::isdigit(word[pqm+2]) && (word[pqm+1] == '-' || word[pqm+1] == '+')) || ( std::isxdigit(word[pqm+2]) && word[pqm+1] == 'x')));
 						bool validPeq = (peq != std::string::npos && (std::isdigit(word[peq+1]) || (std::isdigit(word[peq+2]) && (word[peq+1] == '-' || word[peq+1] == '+')) || ( std::isxdigit(word[peq+2]) && word[peq+1] == 'x')));
 
-						if(validPqm && validPeq && peq < pqm && (abs(pqm - peq) > 1)){ // <cmd>=<int>?<int>
+						if(validPqm && validPeq && peq < pqm && (abs((int)pqm - (int)peq) > 1)){ // <cmd>=<int>?<int>
 							// Dual
 							int64_t val;
 							int64_t val2;
